Utilise un fichier tempraire pour récupérer l'info de la première requête

+++ maquette_get_taxon_info.sh
@@ -34,12 +34,18 @@ cat   TreeFeatures1.json | jq -c '.[] | {sci_name: .sci_name}' | while read json
 	echo " Traitement de $i"
   echo "# taxons = $n, # infos = $ninfo, # images = $nimage"
 	echo "<h1>$i</h1>" >> $resume
+  toto=`mktemp`
+  if [[ ! -f $toto ]]
+  then
+    echo "Erreur pendant la creation du fichier temporaire"
+    exit 1
+  fi
 	touch toto
 	rm toto
 	encoded=$( rawurlencode "$i" )
-	wget -q  -O   toto "https://fr.wikipedia.org/w/api.php?action=query&prop=extracts|pageimages&titles=$encoded&redirects&exintro&piprop=original|thumbnail|name&pithumbsize=400&format=json"
+	wget -q  -O   $toto "https://fr.wikipedia.org/w/api.php?action=query&prop=extracts|pageimages&titles=$encoded&redirects&exintro&piprop=original|thumbnail|name&pithumbsize=400&format=json"
 	# Verifie si l'info existe
-	miss=`cat toto |jq '.query.pages[].missing'`
+	miss=`cat $toto |jq '.query.pages[].missing'`
 	if [ $miss == '""' ]
 	then
 		echo "Pas de données"
@@ -48,14 +54,14 @@ cat   TreeFeatures1.json | jq -c '.[] | {sci_name: .sci_name}' | while read json
 		ninfo=$((ninfo+1))
 		# Recupere l'image
 		#url=`cat toto |jq '.query.pages[].original.source' |sed -e 's/"//g'`
-		url=`cat toto |jq '.query.pages[].thumbnail.source' |sed -e 's/"//g'`
+		url=`cat $toto |jq '.query.pages[].thumbnail.source' |sed -e 's/"//g'`
 		# Recupere le nom  de l'image
 		image_name=`basename $url`
 		echo $url
 		if [ $url  == null ]
 		then
 			echo "Pas d'image"
-			cat toto |jq '.query.pages[].extract' >> $resume
+			cat $toto |jq '.query.pages[].extract' >> $resume
 		else
 			nimage=$((nimage+1))
 			echo "Nom de l'image = $image_name"
@@ -70,7 +76,7 @@ cat   TreeFeatures1.json | jq -c '.[] | {sci_name: .sci_name}' | while read json
 			rm image_info
 			wget -q  -O   image_info "https://www.mediawiki.org/w/api.php?action=query&titles=File:$image_name&prop=imageinfo&iiprop=extmetadata&format=json"
 			# Recupere le texte
-			cat toto |jq '.query.pages[].extract' >> $resume
+			cat $toto |jq '.query.pages[].extract' >> $resume
 			# Ajoute l'image
 			echo "<br><img src=$image_name width = 200px><br>">> $resume
 			# Ajoute l'info sur l'image
@@ -78,4 +84,5 @@ cat   TreeFeatures1.json | jq -c '.[] | {sci_name: .sci_name}' | while read json
 			cat image_info >> $resume
 		fi
 	fi
+  rm $toto
 done
